Refactored all the JSON related functions to a separate file
Created a new file in the dependencies module
called jsonfn.py and added all the JSON functions
to it:

- json_reader
- json_writer
- json_creator
- json_editor

Updated demord.py for the references.
